Refactor service_manager and templates where they belong to
